+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Rewrite of Core/App;
 * Removed the JSONresponse method from Core/Controller;
 * Added a new response utility to set headers and output JSON;
+* Updated Core/Presenter;
+* Renamed Presenter/User to Presenter/Profile;
 
 ## 1.0.5 (Released: 11/04/2017)
 * Added presenters;